🚚rename: [마이페이지] 폴더 구조 변경
- 기존의 조건부 렌더링 대신 페이지 나누는 방식으로 변경

KAN-117
