--- src/app/(route)/(my-page)/_types/myPage.ts
@@ -1,8 +1,0 @@
-import jobData from '../_components/jobTest.json';
-
-export type JobCategory = keyof typeof jobData;
-
-export interface ProfileFormProps {
-  userName: string;
-  userEmail: string;
-}
