+++ src/app/(route)/(my-page)/my-page/check-password/page.tsx
@@ -1,0 +1,9 @@
+import CheckPwd from '../../_components/CheckPwd';
+
+export default function MyPage() {
+  return (
+    <section className="flex justify-center w-full min-h-screen pt-[100px]">
+      <CheckPwd />
+    </section>
+  );
+}
